extract the ability of playing a sound in a channel from sb/index.ts
