+++ src/commands/soundboard/sb/index.ts
@@ -1,10 +1,9 @@
 import DiscordJS from "discord.js";
-import { Readable } from "stream";
 
-import { waitFor, autodisconnect } from "../../../util";
-import soundboard from "../../../db/soundboard";
+import { autodisconnect, playSound } from "../../../util";
 import sbl from "../sbl";
 import Command from "../../command";
+import playSound from "../playSound";
 
 class Soundboard extends Command {
     constructor() {
@@ -14,33 +13,19 @@ class Soundboard extends Command {
     public async execute(message: DiscordJS.Message, ...args: string[]) {
         if (!args[0]) return await sbl.execute(message, ...args);
 
-        if (!message.member?.voice.channel) return message.react("🔇");
         const channel = message.member?.voice.channel;
+        const guildId = message.guild?.id;
 
-        for (let arg of args.slice(0, 10)) {
-            const file = await soundboard.findOne({
-                gid: message.guild?.id,
-                key: arg,
-            });
+        if (!channel || !guildId ) return message.react("🔇");
 
-            if (!file) return message.reply("sound not found");
+        const connection = await channel.join();
+        if (!connection) return message.react("🔇");
+        autodisconnect(message);
 
-            const readable = new Readable({
-                read() {
-                    this.push(file.val);
-                    this.push(null);
-                },
-            });
-
-            const connection = await channel.join();
-            if (!connection) return message.react("🔇");
-            autodisconnect(message);
-            const dispatcher = connection.play(readable);
-
-            await waitFor(dispatcher, "finish");
-
-            dispatcher.end();
-            readable.destroy();
+        for (let sound of args.slice(0, 10)) {
+            await playSound(guildId, sound, connection).catch(e => {
+                return message.reply(e.message);
+            })
         }
 
         return message.react("👍");
